automation/requirements.txt
selenium==2.37.2
behave==1.2.3
pep8
PyHamcrest
